+++ adapter.go
@@ -9,28 +9,28 @@ import (
 	"github.com/aws/aws-lambda-go/lambda"
 )
 
-func handleEvent(ctx context.Context, event events.APIGatewayProxyRequest, handler http.Handler, binaryContentTypes map[string]bool) (events.APIGatewayProxyResponse, error) {
-	r, err := newHTTPRequest(ctx, event)
+func handleEvent(ctx context.Context, event events.APIGatewayProxyRequest, handler http.Handler, opts *Options) (events.APIGatewayProxyResponse, error) {
+	r, err := newHTTPRequest(ctx, event, opts.UseProxyPath)
 	if err != nil {
 		return events.APIGatewayProxyResponse{}, err
 	}
 	w := httptest.NewRecorder()
 	handler.ServeHTTP(w, r)
-	return newAPIGatewayResponse(w, binaryContentTypes)
+	return newAPIGatewayResponse(w, opts.binaryContentTypeMap)
 }
 
+var defaultOptions = &Options{}
+
 // ListenAndServe starts the AWS Lambda runtime (aws-lambda-go lambda.Start) with a given handler.
-// It accepts a slice of content types that should be treated as binary types by the API Gateway.
-// The "*/* value makes algnhsa treat any content type as binary.
-func ListenAndServe(handler http.Handler, binaryContentTypes []string) {
+func ListenAndServe(handler http.Handler, opts *Options) {
 	if handler == nil {
 		handler = http.DefaultServeMux
 	}
-	types := map[string]bool{}
-	for _, contentType := range binaryContentTypes {
-		types[contentType] = true
+	if opts == nil {
+		opts = defaultOptions
 	}
+	opts.setBinaryContentTypeMap()
 	lambda.Start(func(ctx context.Context, event events.APIGatewayProxyRequest) (events.APIGatewayProxyResponse, error) {
-		return handleEvent(ctx, event, handler, types)
+		return handleEvent(ctx, event, handler, opts)
 	})
 }