+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.7] - 2018-06-08
+### Added
+- UseProxyPath option - strips the base path mapping when using a custom domain with API Gateway.
+
+## [0.6] - 2018-05-30
+### Changed
+- Set Host header for requests.
+
 ## [0.5] - 2018-02-05
 ### Added
 - Context support.